// src/routes/type.ts
export type Languages = 
{
    languageID:number,
    name:string;
    place:string;
    native_name:string,
    description:string
}
// ...
export type Content = {
    contentID: number;
    contentText: string;
    englishTranslation: string;
    audioPath: string;
    languageID: number;
    contentTypeID: number;
    isActive: boolean;
    isInDictionary:boolean;
};

export type ContentSyllable = {
    id: number;
    contentId: number;
    syllableText: string;
    audioPath: string;
    orderNumber: number;
};